pyproject.toml
[project]
name = "reachy-claw"
version = "0.1.0"
description = "Low-latency voice assistant for Reachy Mini — powered by sherpa-onnx and OpenClaw"
requires-python = ">=3.10"
dependencies = [
    "reachy-mini>=1.2.0",
    "websockets>=12.0",
    "openai-whisper>=20231117",
    "numpy>=1.24",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
audio = [
    "sounddevice>=0.4.6",
]
sim = [
    "reachy-mini[mujoco]",
]
vision = [
    "reachy-mini[vision]",
]
local-stt = [
    "faster-whisper>=1.0.0",
]
cloud-stt = [
    "openai>=1.0",
]
piper-tts = [
    "piper-tts>=1.2.0",
]
mediapipe-vision = [
    "mediapipe>=0.10.14",
    "opencv-python>=4.8",
]

[project.scripts]
reachy-claw = "reachy_claw.main:main"

[project.entry-points."reachy_mini_apps"]
reachy_claw = "reachy_claw.reachy_app:ClawdReachyMiniApp"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reachy_claw"]

[tool.uv]
override-dependencies = [
    "gstreamer-msvc-runtime>=1.28.1; sys_platform == 'win32'",
    "gstreamer-libs>=1.28.1; sys_platform == 'win32'",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
